feat(ent): 为 AdminPermission 与 AdminRole 添加多对多关联
- 在 AdminPermission 中新增 roles 边及相关查询方法和错误处理
- 更新 AdminPermissionCreate 和 AdminPermissionUpdate 以支持 roles 边的添加、移除及清空
- 在 AdminRole 中新增 permissions、users 和 role_permissions 边及相关查询方法和错误处理
- 更新 AdminRoleCreate，添加 permissions、users 和 role_permissions 边的批量添加支持
- 实现 AdminPermissionQuery 和 AdminRoleQuery 支持边的预加载与条件查询
- 增加关联关系的排序函数支持，优化联合查询的加载逻辑

+++ ent/adminroledept.go
@@ -16,16 +16,17 @@ import (
 type AdminRoleDept struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	// 主键ID
+	ID int64 `json:"id,omitempty"`
 	// 创建时间
 	CreatedAt time.Time `json:"created_at,omitempty"`
 	// 更新时间
 	UpdatedAt time.Time `json:"updated_at,omitempty"`
 	// 删除时间，用于软删除
 	DeletedAt *time.Time `json:"deleted_at,omitempty"`
-	// 角色ID
+	// 角色ID，关联 admin_roles.id
 	RoleID int64 `json:"role_id,omitempty"`
-	// 部门ID
+	// 部门ID，关联 departments.id
 	DeptID       int64 `json:"dept_id,omitempty"`
 	selectValues sql.SelectValues
 }
@@ -59,7 +60,7 @@ func (_m *AdminRoleDept) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			_m.ID = int(value.Int64)
+			_m.ID = int64(value.Int64)
 		case adminroledept.FieldCreatedAt:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field created_at", values[i])
