+++ ent/adminroledept/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.AdminRoleDept {
+func ID(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.AdminRoleDept {
+func IDEQ(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.AdminRoleDept {
+func IDNEQ(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.AdminRoleDept {
+func IDIn(ids ...int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.AdminRoleDept {
+func IDNotIn(ids ...int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.AdminRoleDept {
+func IDGT(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.AdminRoleDept {
+func IDGTE(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.AdminRoleDept {
+func IDLT(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.AdminRoleDept {
+func IDLTE(id int64) predicate.AdminRoleDept {
 	return predicate.AdminRoleDept(sql.FieldLTE(FieldID, id))
 }
